Project.Iterate_Projects: do not process projects twice
Since it automatically replaces extended projects with extending ones,
Iterate_Projects can end up calling the Process callback several times
on the same project (once for the extended project, once for the
extending one). As a result, Enumerate_Mains could return twice the same
main, which in turn introduced errors in the instrumentation of mains.

This is the reason we had crashes in the OA27-059-inlined-separate-cons
testcase in source trace mode. This commit fixes that crash, adding a
protection in Iterate_Projects not to call Process twice on the same
project.

Change-Id: I965fc037d5c2246b5dc5b4da27f91f3f29894542
TN: S918-005