Replace test filters on 7.0.2 by filters on 7.1.2
As we are about to move our check for old toolchains support
from the former to the latter.

Doing this blindly, without actually verifying in all cases
that what was legitimate to filter out with 7.0.2 for technical
reasons remains legit to filter out with 7.1.2, is a bit of an
approximation.  Not a problem for our purpose of backward
compatibility check.

Part of S911-060

Change-Id: I3796728945ac6bf11daf595dc6c717a637941a00