Avoid unnecessary output buffer handling in the x86 disassembler
Disassemblers provide two services: compute instruction length and
disassemble instruction to text. For all disassemblers but the x86 one,
instruction length is constant, so the first service is cheap.

Due to how libopcodes works, the x86 disassembler has to disassemble
an instruction to text in order to compute its length. This is costly,
but there is no way around it. What we can do, however, to reduce the
performance cost is to avoid actually producing a text output (i.e.
writing to an empty output buffer).

This commit implements this, which improves "gnatcov coverage"
performance a lot.

Change-Id: I0191c0fcf6c0f0c77b9c1f9c0060cb515ddfd3c8
TN: P425-025