Traces_Files: refactor read/writes
Trace files reading and writing is currently done through direct calls
to GNAT.OS_Lib.Read/Write. All of them are surrounded with checks that
the amount of bytes read/written is expected, which is error prone. This
commit refactors this part so that only one part of the code does this
check. This makes the rest of the code more readable.

Change-Id: If582cf124c228b82ed2e5952ebad8cd13d467438
TN: SB26-032