Add a sidfile_for interface to LangInfo
This adds a sidfile_for interface to LangInfo for source trace
contexts, similar to scofile_for that we have for bin trace contexts.

This will allow the internal SCOV test driver to retrieve the
name of files expected to hold SCO info for a given source in a similar
fashion from one context or the other.

Part of SB13-022

Change-Id: Idc40778757649577a182b49b7f5f4313a5274895