SCOV.minicheck.build_and_run: make scos take no file extension
In preparation to make SID files processing similar to the handling of
ALI files, this commit changes the semantics of this scos argument to
take extension-less file names for SCOs to load. For now, this is used
only for ALI files, but future changes will be able to use the same
behavior for SID files (produced in the same directory as ALI files)
with a very simple change in SCOV.minicheck.build_and_run.

Change-Id: I3e1da6ac6544a1cf3f92c2d78c990417aaf808a3
TN: SB13-022