In-place instrumentation of expression functions for MC/DC
When instrumenting expression functions for MC/DC, replacing
them with generic instantiations raises tricky issues with freezing
and access before elaboration (because instantiating a generic with
the expression function as a formal causes freezing of the parameter
types). Instead use a different instrumentation patterns only
introducing a new expression function in-line (in a wrapper package).

Fixes regressions in previous change for S725-026

Change-Id: I1b04037c82e941a32cb6040d2c640e199e3ea878