Rename --dump-method to --dump-trigger
This option determines *when* to dump coverage buffers (at the end of
the main subprogram? when exit() is called? manually?), so "trigger" is
better suited than "main". *How* the dump is done (i.e. the dump method)
is always the same for now: calling
GNATcov_RTS.Traces.Output.Write_Trace_File.

An upcoming change will (re)introduce --dump-method to make the
procedure to call for the dump customizeable.

Change-Id: I7b8e13cb1f392b034c32d201b2325f9fd16278b3
TN: T110-022