Fix STR/genrest.py retrieval of common compilation switches
STR/genrest.py calls control.BUILDER.SCOV_CARGS passing
it a dictionary representation of the testsuite command
line options (retrieved from a json dump produced when the
testsuite ran) but SCOV_CARGS expects an instance of a
command-line-option kind of class.

SCOV_CARGS just interrogates basic properties of the
<options> input, e.g. fetch the string value associated
with such particular option, so this change arranges
for genrest to pass a namedtuple object constructed
from the dictionary.

Take the opportunity to add a comment describing the role
of the <options> argument to SCOV_CARGS.

Part of T313-036 (gnatcov qualification for qgen)

Change-Id: I7f2ea0e85953f183dba623a98ed69be775d32e4a