Add trace_mode to the options dictionary dumped for qualification
As the trace_mode is needed afterwards to re-infer the common
compilation flags at report production time.

This logic will probably need to be refactored at some point (so
we just fetch the sets of options used), but better get the original
design to work again. It has merits of its own.

Part of T313-036 (gnatcov qualification for qgen)

Change-Id: Ic438914ab3be1ed112920e75297d0e54382c72d2