Reword the Tool Overview section of PLANS
To remove the inaccurate statement that the tool
(always) operates on code that can be deployed on
the target hardware. Not true in environments
where coverage analysis is performed on native host
programs and representativeness is acknowledged from
equality of test results.

Also adjust the reference to the foundation publications
to advertise them as describing the tool's foundations,
not detailed characterstics of the tool as it is today.

Part of T419-005

Change-Id: Idf8d0db5099d912a58bf4d75e84f806b43b126f2