Add a GprFacilities section to Qualif/Common
This part holds the bits describing the use of general
project file facilities such as attributes for switches
or the support of --subdirs.

The part with everything related to stating units of
interest will be added separately.

Initial set of tests copied from tests/GPR, reorganized
and augmented with requirement and testcase descriptions.

The reorganization is essentially about directory renamings
to be more descriptive and better fit in testing strategy sections
of requirements.

Part of T317-055

Change-Id: Id8a31ba2b439ceb4042161866f06a91ceb10881e