Refactor QM TestCasesImporter to support other subdirs of "Common"
This change factorizes the global TestCase import processing,
and in particular the code which produces PDF outputs with the
introduction of a tc_pdf_for method.

Previously, there were several blocks repeating the same sequence
with only very minor differences which could easily be synthesized
(e.g.  depending on wether we are outputing the first chapter
of the whole set).

The main goal is to facilitate the forthcoming addition of
subdirs in the Common part, so it can now be done by just adding
a subdir name to a list instead of replicating complete
sequences of instructions.

Part of T317-055

Change-Id: I3630f3ed347806693bacc8fc37de1354f75cee32