Relax constraints on expr determining category from dirname
To determine the stmt/decision/mcdc category of a SCOV
test from it's directory location, we used to match the dirname
against /<category>/ or /<catagory>$

/<category> is enough in practice and allows more flexibility
in the selection of names. This is useful when creating multiple
trees of tests and we don't need/want to introduce an artificial
intermediate directory just to satisfy the match.

We can now create, for example:

  .../stmt-series1
  .../stmt-series2

instead of

  .../stmt/series1
          /series2

This is useful in the context of qualification material
where every intermediate layer requires a description and
adds documentation overhead.

Part of T317-055

Change-Id: I649004ff6fd96bfbc5b3f74545a5295f74e9a909