Add a GPR section to Qualif/Common/UnitsOfInterest
This adds the material validating the use of project files to
specify units of interest. Taken and heavily adapted from the subset
of tests/GPR/Coverage that dealt specifically with units selection
(vs the possibility to control switches, support for --subdirs etc).

All the tests that were in tests/ are in this new Qualif/
section one way or another. The reorg removes redundancies and
distributes the tests in a tree much easier to use as the basis
of a testing strategy description.

Part of T317-055

Change-Id: Ia82d3b11d201a4b7365d8e9816dad8cf7104452f