Add support for subdirs and xvars to GPRswitches
This allows easier testing of the corresponding
facilities using the SCOV helper classes and facilitates
the handling of GPR tests sharing project files in
a common subdir, as we can more easily arrange for
each test to use a unique object dir for itself.

The bulk of the change consists in identifying
a category of switches that need to be passed at build
time as well (in addition to gnatcoverage invocations).

Part of T317-055

Change-Id: I9eda701db29a59eeb8b40d5f8a3161f85685be82