Tighten [] vs None for CovControl units attributes
Passing None or [] for arguments like units_in differs
significantly as it expresses either "I don't want an attribute
at all" or "I want an attribute with an empty list value".

Clarify this in the spec comments and account for this in the
predicate determining if a project file is needed to convey coverage
attributes.

Part of T317-055

Change-Id: Idd6d928ea232a59a88c3085c94d9c0f6e7b8eff8