Add a mechanism allowing tests to import a test_support module
For starters, when a test_support.py file is available in the
test parent dir.

This is useful to allow the use by individual tests of common
facilities meaningful to only a specific group of tests and which
don't belong to our more general SUITE or SCOV modules.

Part of T317-055

Change-Id: I465f95140b3f13a01df8889cb4b23b4bccc26ff2