Simplify the doc guidance regarding options for mcdc
Remove the Trace Control for MCDC section, too much potentially
confusing detail for a user's guide. Synthesize the important points
in the doc of --level for gnatcov run.

This also removes a part of the doc where we were duplicating
information about how to specify units of interest which had turned
obsolete after the recent change of default wrt recursiveness.

Part of S528-004

Change-Id: Id2d10a03c03f810c5731f8fa5cf7e5a167dbecb5