Qualif/Common/GprFacilities/ConfigAttr/Target: fix on native targets
During the testsuite configuration step, always pass --target to
gprbuild (i.e. for cross *and* native targets).

This fixes the "default GPRbuild target" (for instance x86_64-linux)
versus "default gnatpython target" (x86_64-pc-linux-gnu) discrepancy.
Note that we used not to do this because we used to run the testsuite on
x86_64-linux boxes while testing against a x86-linux toolchain without
using the testsuite --build option. Nowadays, this option is always
used, so this configuration works even if we pass --target to gprbuild.

Change-Id: I564a7c64d1ef9f6ecdc511c807f887f12bca9200
TN: T706-021