Improve Qualified Interface description in PLANS
Merge a few individual changes done in a qualification project
context, that "conflict" (in git terms, not conceptually) with
changes issued on our development trunk to incorportate support
for project files.

Clarify that --target is only for cross configuration.
Clarify that --scos is only for mcdc
Rework text arrangements to avoid misformatting

T419-005, part of T721-013, comments line#21 and #24 on PLANS

Change-Id: I71aef1afca36d8597a5c91aaccfb906e7cac160f