SC_Obligations: fix duplicate loading of BDD nodes from checkpoints
BDD nodes (descriptions of decision structures) are materialized in
checkpoints as one big table, and each decision entry in SCOs points to
a range of BDD nodes.

The way checkpoint loading was implemented so far imported the full
table of BDD nodes for every compilation unit described in the
checkpoint, creating lots of duplicates, and most importantly, wasting
heap space, sometimes leading to out-of-memory conditions.

This commit changes this checkpoint loading process to import BDD nodes
only "on demand": when going through the decisions that reference them,
thus removing the duplication.

Change-Id: I385088942e3c9bc09b6aabd07981acdc5fa7b031
TN: T731-036