Replace the XML report DTD with a XML Schema
XML Schemas are considered as the evolution of DTD and enable stricter
checking. This commit also updates the report "grammar" according to the
recent XML report extensions, and the XML report generation to refer to
the XML Schema file (*.xsd), copying it next to the generated report.

For the record, the XML validation process itself is not automatically
tested as of today: our testing infrastructure currently does not
provides the required tools. This lack of testing led to the
obsolescence of the DTD, so the middle term plan is to do the necessary
work in the testing infrastructure and then start reasonable automatic
testing for this feature.

The XML Schema this commit introduces has been tested on a simple XML
report using the Python lxml library (https://lxml.de/):

    import lxml.etree

    # Read the XML document and expand the XInclude directives
    with open("index.xml") as f:
        doc = lxml.etree.parse(f)
    doc.xinclude()

    # Run the XML Schema validation
    with open("gnatcov-xml-report.xsd") as f:
        xsd = lxml.etree.XMLSchema(lxml.etree.parse(f))
    xsd.assertValid(doc)

Change-Id: Iffd24a1194d0c999600902f63ba445631d37f1b8
Depends-On: I616228e80fe5c3642563abf0e37906e5b639cf6c
TN: T803-013