Create our own data type to represent pragmas
Currently, the compilation of sc_obligations.ads is very tied to the
version of GNAT, as this source contains an aggregate to initialize an
array indexed by Pragma_Id (from gnat_util, and thus from GNAT itself).
This means that additions/deletions of pragmas in GNAT results in build
failures for GNATcoverage.

This makes it impossible to build GNATcoverage without the exact
matching version of GNAT. In addition, this sometimes means removing
support for a pragma from GNATcoverage, which is not something we want
to do (for backwards compatibility with previous version of GNAT).

This commit introduce our own Pragma_Id type, which we can extend as
much as we want, to cover all known pragmas. A simple testcase using
gnat_util will then check that gnatcov knows at least about all pragmas
from gnat_util.

Change-Id: I9f38b0f0bf8b4559948dceda677bc3ccf6145a16
TN: T629-012