Update gnatcov_rts for Ada95 & old toolchain compliance
We need to allow compilation of the coverage runtime
with Ada95 toolchains, potentially old. We have an eval
about to start for use with GNAT 5.04a.

This change addresses to minor issues in the gnatcov_rts
scope (not rts_full).

Remove use of 'Mod, which is only available starting with Ada 2005.

Remove "constant" qualifier on local array declaration with
an address clause and an Import pragma, which GNAT 5.04 wrongly
rejects as an incorrect deferred constant declaration.

Part of T811-008

Change-Id: I60f50b25fe9e89df42799972564d82ddc94dcb15