Instrument.Find_Units: do not trigger PLE in Libadalang
PLE (Populate_Lexical_Env) is an analysis unit operation that
initializes semantic data in Libadalang, and in particular which fetches
the closure of dependencies. The P_Fully_Qualified_Name property
automatically triggers PLE, which is not necessary for Find_Units's
operations. As an optimization this commit switches to the use of the
P_Syntactic_Fully_Qualified_Name property, which is sufficient for our
needs, and does not trigger PLE.

Change-Id: Ided12c8ca1e369d4f4af640fe5cc13fa2f47cfdb
TN: T824-029