Checkpoints loading: fix remapping of BDD node IDs
First, some context about the loading process for checkpoints (or SID
files): checkpoint-related data is organized both in gnatcov and in
checkpoint files as descriptor tables, so a reference to a descriptor in
such a table is just an index.

As a result, loading a checkpoint in gnatcov's internal tables is
conceptually made up of two basic operations: concatenate the checkpoint
file's tables to gnatcov's in-memory ones, and "relocate" them, i.e.
update the indexes they contain to go from "in-file indexes" to
"in-memory indexes".

When tables are loaded "all at once", this index update is simple: just
add the current size of the table to the indexes that refer to the table
that is loaded from the checkpoint file.

When tables are loaded "chunk by chunk" (which is the case for BDD node
tables), the index update is still an addition, but we must consider the
size of the internal table *right before* the current chunk is loaded.

This commit fixes BDD node tables loading to consider the internal
gnatcov table for BDD nodes right before loading a BDD (i.e. a subset of
nodes in the BDD nodes table). The previous implementation considered
the size when starting to load a unit ("CU" in SC_Obligations) whereas
a unit can contain several BDDs.

Change-Id: I21d44d273dfa7c91203a64098f3129c00317b7c0
TN: T916-046