T921-012 integrating disassembling support for aaarch64-elf
The support for the aarc64-elf platform.
The main modification is the new partial disassembler for aarch64,
that tells which instruction is a branch and what kind of branch.
The specification for the instruction codes is linked in
disa_aarch64.adb, in the Get_Insn_Properties procedure which makes
use of it.

Change-Id: Ia936fd137f826943c7ef2db8be25eee0fca3957f
TN: T921-012