Refine Qualif material for if-expressions
Account for their relevance to decision coverage, in the Core
DC requirements first and adjusting the MCDC requirements consequently.

Add DC tests, part of the Topologies family where we
cover the various contexts where decisions are to be expected.

Refine the Value and Flip subcases more generally, which exercise
Unary expressions. For such expressions, dominance in conditional
contructs controlling statements allows us to tell the difference
between dT- and dF- with both binary and source traces.

We have no way to make this distinctinction with binary traces on
IF-expressions. Add dedicated "ifx-eval" expectation anchors for this.

T310-018

Change-Id: I3e12884a864d365b606fe0a948f990358b6d2600