Support for Angular GNATquilt version
Adapted the previous output format for GNATquilt (Annotations.Dynamic_HTML)
for the new Angular version.
The underlying data model is the same:
  - JSON data for the whole report, to get aggregated information.
  - JSON data for each file, to get detailed information.

Main changes are the generation of the HTML and the format of each generated
script.
The HTML inlines different scripts, as the framework used is not the same
(Google Closure for gnatquilt 0, Angular for GNATquilt 1).
The script generated for each files, where the JSON data is inlined also
have a different format.

For recall, in GNATquilt 0, for each source file there was a script which
was just a call to a function of the like:
gnatcov.load_hunk (JSON_DATA_FOR_FILE);
The aggregated data of the report was inlined in the index.HTML as a similar
call to a different function.

Now, a global variable is filled every time a new file is loaded in the HTML
report.
The script is then as such:
REPORT [FILENAME] = JSON_DATA_FOR_FILE
There is a report script, that could probably be inlined as well in
the HTML, which declares this global variable and fill data for the report.

The needed scripts and CSS files will have to be present in the angular
folder of the library folder (check for the existence of that folder to
know if plugin is installed or not, not for individual JS and CSS files).

Change-Id: If402bb9dc0b7c188e125daf2e050a6b319643eae
TN: TA14-003