gnatcov_rts_full: rework dates handling for Ada 95
We are aiming at compatibility with older toolchains, down to
Gnat 5.04a and thus limited to Ada 95. Some of the functions available
in Ada 2005 in the Ada.Calendar package are not in Ada 95.
This means we have to find an alternative for the generation of timestamps,
and not use Ada.Calendar.Time which did not have conversion operations to
Unix time back then.

Now using the standard C function time as a default timestamp, and users
will have to provide timestamps as such when calling Write_Output_File
of the runtime library.

Change-Id: I69b1264e3fd6456a60e080197390032dfaceaf09
TN: T811-008