Instrument: fix MC/DC handling with expression function
This commit fixes another case of wrong assumptions regarding the index
of declarations in rewritten lists of declarations: the index of a node
before rewriting can be very different from the index of the same node
after rewriting. For this reason, we must re-compute a node's index when
attempting to remove it from the declaration list it belongs to.

Change-Id: Ic3b7c39e4856317eeff5efd070e29a2ed4763687
TN: TB13-020