Instrument: fix naming issue in mains for automatic dumps
This commit reworks the way mains are modified to to add
an autodump for the coverage buffers.

Add_Auto_Dump_Buffers moves all the declarations and statements
from the main in a nested declare block to be able to insert calls
to register trace dumping before the main's original code, or
dump the traces at the end of the main.

This commits names the inserted declare block, and replaces
qualified names of the kind "Main.Foo" by "Main.Original_Main.Foo".

TN: TB06-028
Change-Id: I9e7d0e20e2dca07456f50e55f1e6dd732254f83a