Prevent optimization of Witness calls with GNAT5.04a1
The following:
if False then
   null;
elsif False then
  null;

will be instrumented as followed:

Witness(1);
if False then
   Witness(2);
   null;
elsif Witness(3) and then False then
  Witness(4);
  null;

When compiling with GNAT5.04, this gives us, with simplifications
(-gnatG), the following:

Witness(1);
if Witness(3) and then False then
  Witness(4);
  null;

Note how the first alternative is optimized out, as the program will
never enter it, but not the 2nd one, which will not be entered either.
The optimization of the first alternative (if False then ...) is
valid, since the compiler statically knows it is dead code. The call
to Witness(1) is preserved, so traces will show that "if False" is
covered.

However, even if the second alternative is preserved according to
-gnatG output, it is actually optimized out in machine code. This is
also valid, since the compiler knows that whatever Witness(4) returns,
the decision will evaluate to False, and since the Witness function is
pure (no side effect *allowed*, even though we cheat on that part),
the compiler is allowed to elide the call. Unfortunately, this prevents
traces from showing that "elsif False" is covered. This is what this
patch tries to address.

To prevent Witness (possibly executed) calls from being optimized out,
modifying the instrumentation scheme of elsif constructs as follows
seems to be enough in practice.

Changing the resulting boolean expression from

elsif Witness and then Expr --  Witness returns True
to
elsif Witness or else Expr  --  Witness returns False

results in 2 things:

- The decision can be simplified to True, if we have for example
  "Witness or else True". This means further alternatives (following
  elsifs) can be removed. It will not be a problem, because those
  correspond to actually unreachable code. We only care about Witness
  calls in executed (even symbolically) code.
  In that case, the Witness call seems to be preserved, even though
  it would be valid for the compiler to remove it.
- The decision can no longer be simplified to False, which means the
  elsif alternative will not be optimized out. The Witness call thus
  makes its way to machine code, even though it could be omitted in
  theory.

Change-Id: If122e835e3812faf5f853f2b5ecc6ce365c94a33
TN: T811-008