Instrument: emit declaration for self-referenced expression functions
Previous changes for this ticket disabled the emission of a declaration
for the instrumentation of expression functions. For instance, given the
following expression function in user code:

    function F return T is (...);

We used to emit the following:

    function F return T;
    package Wrapper is
       function F return T is (...);
    end Wrapper;
    function F return T is
      (case [witness call] is others => Wrapper.F);

And since the previous changes, we now emit the same, without the first
"function F" declaration. This is unfortunately not correct when the
expression in function F references F itself.

This commit refines the previous change so that the "function F"
declaration is not omitted when F references itself.

Change-Id: I9736b3cdca647a2b1980d2ce750fc398fc3ee064
TN: TB20-035
(cherry picked from commit d801fcc73bf23e2cbf2752b312b9814183b01752)