Instrument: tag statements as having code only once
While with binary trace, "statement has code" means "there is machine
code whose debug information relates to a sloc in that statement",
things are very different with source traces, which don't involve
analysis of machine code.

For source traces, we consider that a statement has code as soon as we
inserted a Witness call for that statement, i.e. as soon as we allocated
one bit for this statement in its unit's statement buffer. This means
that we don't need to wait for the processing of a source trace
referencing a statement in order to tag that statement as "no code":
tagging it when loading the SID file that describes this statement is
just as fine.

As a bonus, this is also much more performant, as we tag each statement
as having code exactly once, instead of once per source trace
referencing that statement. With a lot of sources traces, this
optimization can reduce ten times the time to execute "gnatcov
coverage".

Change-Id: Ib734897e8e04bb0769e66e73e4655bb64d9adce6
TN: TC04-045
(cherry picked from commit 759109da443305a110b6e2772fe661e7741ad94f)