Change gpr configuration for 5.04a1
Having C++ as a language in the suite.cgpr config file forces the use of
g++ as linker. This results in incompatibilities with the lib_gcc.so.1
provided by the GNAT 5.04a1 toolchain.

To avoid that, C++ is removed from the config file when running the
testsuite with GNAT 5.04a1.

Also applied a minor refactoring when computing toolchain discriminants,
and fixed a wrong comment in the same section.

Change-Id: I86e38cb1894f04b1f9508b0c9f85df17b40c9033
TN: T811-008