GNATcov_RTS.Traces.Output.Files: default filename in specified directory
This commit extends the
GNATcov_RTS.Traces.Output.Files.Default_Trace_Filename function so that
the environment variable can contain the name of a directory. In that
case, the source trace file is created in that directory *and* using the
default behavior regarding its basename.

This will be useful for testsuites to just tell the runtime to create
non-conflicting source trace files in a single directory.

Change-Id: Iaee167eb77305f6ad714d4756b4f7165fb6e737c
Depends-On: I985da90f1c45d7dbf018ab1cc744e2b6a6dc672b
TN: TC04-034
(cherry picked from commit 727a7686fc8bc327d9db99195385d539d585130d)