Altrun/trace32: give more time for the t32 server to start
The first test of the testsuite is always failing because it can't
connect to the trace32 remote interface. This could be due to an
insufficient delay between starting the t32 app and running the first
test. The increase in delay should not be a problem given that it will
only run once and is rather insignifficant compared to the ~5 hours it
takes for the testsuite to run.

Change-Id: I2da83a39bb2673bffd80d0d5f2cb12f5d25d97f4
TN: TA27-037