Traces_elf: Fix Find_Address_Info
This fixes a wrong assumption concerning the ordering of Address_Info
which lead Find_Address_Info to sometimes return an Address_Info with
bounds not containing the desired address.

This fix only works under the assumption that given two Address_Info
then they are either disjoint or one contains the other.

If it is found that this assumption does not hold, then we may need
to revisit this function.

TN: S320-042
Change-Id: Ice9701ae14fe4468ac86cfde96574199ed3ff47d