Prepare gnatcov trace_adapters/valgrind to valgrind 3.16
Account for a minor internal change corresponding to the
move from --vex-guest-chase-thresh=<int> to --vex-guest-chase=<yes/no>
between 3.14 and 3.16. Use conditional compilation to support
the two versions.

Robustify our Makefile wrt the possible presence of a <prefix>/libexec
in addition to <prefix>/lib or <prefix>/lib64.

Part of T730-028 (upgrade valgrind for ubuntu 20.04)

Change-Id: Iab21962235736044b8727433e176a1c1e0d406f2