Complete the instrumentation doc on the control of trace filenames
Allocate a standalone section on the control of trace filenames,
part of the "Instrumenting programs" section. Rework the section where
that info was located.

The previous info has just became incomplete with TC04-034 and should
now legitimately be introduced as part of the intrumentation description
as quite a bit of control is now offered by the instrumentation command
line.

Part of TC04-034 (support for trace filename variations)

Change-Id: If1d488347e71f88025b433117d1aacdd45044745