Introduce a gnatcov driver
This commit creates a "gnatcov driver" that runs the correct gnatcov
(32-bit or 64-bit) depending on the selected (or inferred) target. It
also does the appropriate build system changes to build both "bits"
specific entry points and to install them in the expected place, and the
appropriate testsuite changes for 32-bit runs to work correctly.

Change-Id: Id273476135cae0fd4059527e307d0bfb7c479874
Depends-On: Ib5d55733617e28e69e9369a814a71464a52ae695
TN: U129-011