Refactor Ada instrumentation packages
This commit merges the Instrument.Sources (instrument-sources.ad*) and
Instrument.Tree (instrument-tree.ad*) packages, that were designed for
instrumentation of an Ada source file.
It also moves some of the Ada specific instrumentation code that was in
the Instrument and Instrument.Common packages (instrument.ad*,
instrument-common.ad*), down to the package specific to Ada
instrumentation.

Code in Ada specific packages (Instrument.Sources and Instrument.Tree)
that can be shared with other languages instrumentation process is
moved up to the Instrument.Common package.

This is preliminary work for C instrumentation.

Change-Id: I4804eba07522b11839fe856c0cce540fd8391cce
TN: U208-020