Instrument: fix default source trace dump settings
When the --dump-channel is not passed (and thus that bin-file,
the default, is used), the default trace filename should be complex
(a prefix plus several timestamps), unless --dump-filename-simple is
passed.

Because of a wrong default initialization, not passing --dump-channel
used to automatically enable --dump-filename-simple. This commit fixes
this inconsistency.

Change-Id: I252458182ee7e8c762b090b2c2c530de368e4f84
TN: U210-019