Consistently use Support_Files.Gnatcov_Command_Name
... instead of Ada.Command_Line.Command_Name. The former returns the
command that the user typed, which is what we want to use in error
messages (we don't want the absolute path to the bits-specific entry
point instead, which Ada.Command_Line generally returns).

Also restrict the emission of the "Could not spawn
/path/to/bits-specific-entry-point" message to verbose mode, as this is
just noise most of the time. Putting it in verbose mode will be useful
when investigating a potential failure in this area.

This change restores the corresponding behaviors of gnatcov before the
introduction of the gnatcov driver, and thus fixes the corresponding
user experience regressions.

Change-Id: I99fc53e1c05df7fbc83ffba81c8e5ee5733d56e4
TN: U217-018