traces_elf.adb: Simplify processing of symbols for PE executables
In such executables we only have the start address of a symbol,
so it is considered that a symbol's address range spans from its
start address until the next symbol's start address. With such an
assumption, there is no need to keep track of which symbol is top
level or not, as they all are.

Removing the insertion with top level update also has the benefit
of removing checks on the symbols Address_Info when the end address
was not yet specified.

Change-Id: I8afdee6264a5f130d90c4de1bae402f3c32cf271
TN: S320-042