Add a mechanism to keep track of context to report internal errors
This commit introduces a simple mechanism so that gnatcov keeps track of
what it is doing so that it has some basic information to contextualize
the internal errors (by nature unexpected) reported to users.

This also uses this new mechanism to note which node the instrumenter is
processing to contextualize internal errors in the instrumentation
process.

Finally, this also introduces a mechanism to trigger artificial errors
at specific points through an environment variable, for testing
purposes: internal errors are by definition bound to be fixed, so we
need artificial errors to test their reporting facility.

Change-Id: Ifde4d4716bac73b5ea8d20eab33c0efc4d95e36f
TN: TC11-052