SCOV.instr: by default, fail if there are warnings in the output
Upcoming changes will add warnings to the output of "gnatcov
instrument" to warn users when the instrumented code is likely to
trigger a compiler bug.

These warnings are supposed to appear only when instrumenting very
specific constructs: we want to check in the testsuite that they don't
appear otherwise. So prevent them by default.

Change-Id: I3058c1b47038b386b55a00bbe25a50bef1dd95c6
TN: TB30-009