files_table.adb: Rebase files before inserting them in table
Before inserting a source file whose full path is known, check
whether it really exists on disk, and if not try to locate it
with the source-rebase/source-search information we have.
This avoids the duplication of file entries of a same file
coming from checkpoints (or traces) produced in different machines
for instance.

Change-Id: I8494132986fdafbabaedb1ab4e7014b240a0e04d
TN: U204-029