Avoid crashing on wrong dominance marker sloc
GNATcoverage crashes on a constraint error when a SCO associated to a
dominance marker is not found in the SCO decision map (because of wrong
computed source location or whatever).
It will now emit a warning in those cases, and discard dominance
information when that happens.

Change-Id: I457ca0e3676fb9cd089c396bdd67393cc6a363df
TN: U219-010
(cherry picked from commit 3daa91312166b146e4925592b7bee9e18b62debc)