Refactor Compilation_Unit_Name type
In order to support other (than unit-based) languages with GNATcoverage
source instrumentation, we need to extend the high level instrumentation
scheme and particularly how compilation units are described.

This commit extends the definition of the Compilation_Unit_Name type so
that it can be used for file-based languages (notably C).

It also introduces very basic support for the C language, to make use of
the extended type.

Change-Id: Icc670619e5c6e21bc8db4310c086a0a0f3fb274f
TN: U208-020