Fix GNATcoverage XML report
This commit fixes the column_end field as its value was always equal to
to the column_begin field.

It also removes the implicit values for column_begin and column_end, and
outputs those fields even when they are equal, which was not the case
before.

The example given in annotations-xml.ads is updated to fit GNATcoverage
current XML report.

Change-Id: I8ee8add4f081e78f9e38e5ff74565dca07b67b96
TN: U317-004