--scos/--sid: Warn about incorrect combination of options
After this commit gnatcov will reject the use of traces
(or checkpoints) of different kinds, unless the internal
option --mix-trace-kind is passed, in which case it will
only warn about the mixing of trace kinds. This introduces
a backwards-incompatible change in the checkpoint format.

Gnatcov will also warn about the use of --sid in binary
trace mode and --scos in source trace mode.

This also updates the doc to make it a bit more explicit
which option can be used with which kind of trace.

Change-Id: If555aadea0152c8a03c6433226eed1a3825f4823
TN: T909-026