Add gnatcov handling of pragma Gnat_Annotate
Recently introduced in GNAT, synonym of pragma Annotate.

This change just adds the pragma name to the list of
pragmas known to gnatcov and arranges to handle it the
same way as pragma Annotate (not generating code on its
own, so not to be considered as a statement).

Change-Id: I2565fc3701c6728ab6e397f816aa5cdd92e29d05
TN: U313-005