Instrument_Ada_Unit.adb: Accept QuantifiedExpr not in ParenExpr
Gnatcov assumes that every quantified expression is enclosed in parenthesis.
This is not always the case, when the parenthesis are part of
another node, such as a subprogram call or an assertion/loop-invariant.

This commit makes gnatcov instrument correctly quantified expressions
when their parent node is not a ParenExpr.

Change-Id: Ie5f100f329e66286712df2f65e212ace6747d594
TN: U526-032