Correctly handle aspects when emitting a decl for expression functions
During instrumentation, when emitting a declaration for an expression
function (case of self-referecning expression functions), gnatcov
moves the aspects of the original expression function to the
declaration, but then immediatly removes them to include them in the
augmented expression function.

This change makes it so that aspects are only added to the augmented
expression function when no declaration was emitted for the original
expression function.

Change-Id: I999e7a447547c9f8a98f7cb8e2d029724ec7947f
TN: U603-021