Disable instrumentation of assert-like pragmas by default
Assertion pragmas are currently always instrumented and thus 
generate SCOs that are always violated. This change disables by 
default the instrumentation of assert-like pragmas (assert, 
pre/post, check, loop invariant), to be more in line with how the 
pre/post aspects are processed (they are simply not instrumented).

It also adds the loop_invariant to the list of pragmas for which
decisions scos are ignored with level stmt+decision to unify
the handling of such assert-like pragmas.

Change-Id: I6bade5db7139d4a013527ff31cdf5eb97d319267
TN: U528-022