gnatcov instrument: fix instrumentation of elsif statements
Elsif statements with a decision being an if expression, or a quantified
expression were not rightly instrumented.

As a reminder, we instrument such statements by enhancing the condition
with a witness call (that always returns false), e.g.

"elsif C" becomes "elsif Witness or else C"

If the condition C is a complex decision, and does not have an or else
operator at the top level, it needs to be rewrapped in parenthesis. The
case where C is an if expression, or a quantified expression was not
considered.

Also refactor the elsif instrumentation scheme to make it more generic.

Change-Id: I47d47f2397e40a9c8fc5746e513421c39a21a40b
TN: U622-031