Rework gnatcov doc on --ignore-source-files
* Source Coverage chapter:

In the option summary for --ignore-source-files, refer to the
section dedicated to the description of the option and related
attributes (within the chapter on units of interest).

Move the text describing the handling of subunits
to this section.

* Units of Interest chapter:

Rework the description of --ignored-source-files to ...

- Present it more explicitly as a possible alternative to
  the project file attributes,

- Provide example command lines and response file

- Incorporate the partial information which previously
  was in the "Ada subunits" section.

Part of U624-012 (Improve gnatcov doc on --ignore-source-files)

Change-Id: I803c42d6cadfe2a33624c5a9b154c00262556bfc